Match latest TodoMVC app specification
While we don't use bower for dependencies, we can certainly load our
CSS from the same directory as everybody else to make this easy to keep
up to date.

Otherwise, I also updated the index file to bring in changes made to the
mopped version deployed on the official site.